feat: bind data into the ui and design layouts for different states of the request

+++ app/src/main/java/com/example/tomorrowweather/model/response/TimeStamp.kt
@@ -16,7 +16,7 @@ data class TimeStamp(
     @SerializedName("visibility")
     val visibility: Int?,
     @SerializedName("pop")
-    val probability: Int?,
+    val probability: Double?,
     @SerializedName("sys")
     val timeOfTheDay: DayPart?,
     @SerializedName("dt_txt")
